+++ docs/changelog.md
@@ -9,12 +9,14 @@ will continue to work with version 0.6.1, but may break with version 0.7.0.
 ## Unreleased
 
 - BREAKING: Drop Python 3.6 support.
+- BREAKING: Require {doc}`importlib-metadata<importlib_metadata:index>` 3.6 (was 3.4).
 - BREAKING: Drop the `cernml.coi.__version__` attribute. To query the COI version, use instead {mod}`importlib_metadata`. (With Python 3.8+, this is in the standard library as {mod}`importlib.metadata`.)
 - BREAKING: Remove `PascalPase`-style members of {class}`~cernml.coi.Machine`. Use the `SCREAMING_SNAKE_CASE`-style members intead.
 - BREAKING: Remove `cernml.coi.unstable.japc_utils`. It is now provided by {doc}`cernml-coi-utils<utils:index>` as {mod}`cernml.japc_utils`.
 - BREAKING: Remove `cernml.coi.unstable.renderer` and `cernml.coi.mpl_utils`. Both are now provided by {doc}`cernml-coi-utils<utils:index>`'s {mod}`cernml.mpl_utils`.
 - BREAKING: Remove `cernml.coi.unstable.cancellation`. The module is now available as {mod}`cernml.coi.cancellation`.
 - BREAKING: Remove `cernml.coi.unstable`. The module is now empty.
+- ADD: Support for {doc}`importlib-metadata<importlib_metadata:index>` 4.
 
 ## v0.7.5